+++ src/ui/circle/circle-level-drawer.js
@@ -32,7 +32,6 @@ function drawCircleLevel(typeOfMenu, E, mangle, mradius, buttonsToShow, circleRa
         var outlineColorRgb = getTextColorForBackground(colorForButton, 1.0);
         var outlineColor = `rgba(${outlineColorRgb.red}, ${outlineColorRgb.green}, ${outlineColorRgb.blue}, 0.5)`;
 
-        // if (selectedButtons[level] == i) {
         if (selectedButtons[level] == i || preselectedButtons[level] == i) {
             /// Segment is hovered